+++ examples/04-write-to-persistent/server.c
@@ -12,6 +12,7 @@
 #include <stdlib.h>
 #include <stdio.h>
 #include "common-conn.h"
+#include "common-hello.h"
 #include "common-map_file_with_signature_check.h"
 #include "common-pmem_map_file.h"
 
@@ -63,7 +64,7 @@ main(int argc, char *argv[])
 		 * All of the space under the offset is intended for
 		 * the string contents. Space is assumed to be at least 1 KiB.
 		 */
-		ret = common_pmem_map_file_with_signature_check(pmem_path, KILOBYTE, &mem);
+		ret = common_pmem_map_file_with_signature_check(pmem_path, HELLO_STR_SIZE, &mem);
 		if (ret)
 			goto err_free;
 	}
@@ -72,11 +73,12 @@ main(int argc, char *argv[])
 	/* if no pmem support or it is not provided */
 	if (mem.mr_ptr == NULL) {
 		(void) fprintf(stderr, NO_PMEM_MSG);
-		mem.mr_ptr = malloc_aligned(KILOBYTE);
+		mem.mr_ptr = malloc_aligned(HELLO_STR_SIZE);
 		if (mem.mr_ptr == NULL)
 			return -1;
 
-		mem.mr_size = KILOBYTE;
+		mem.mr_size = HELLO_STR_SIZE;
+		mem.data_offset = 0;
 	}
 
 	/* RPMA resources */